create TestCaseMiddleAndBottomPanelFactory
